refactor(knockdog): searchState fsm 구조로 리팩터링
